requirements.txt
# Production
dj-database-url~=0.5.0
djangorestframework~=3.12.4
django~=3.2.11
psycopg2~=2.9.1
Pillow~=9.0.0

# Development Only
pytest~=6.2.5
factory-boy~=3.2.0
pytest-cov~=2.12.1
pytest-django~=4.4.0
pytest-lazy-fixture~=0.6.3
pytest-mock~=3.6.1
ipdb
